+++ rocketpy/simulation/flight.py
@@ -1692,6 +1692,27 @@ class Flight:
         ax, ay, az = K @ Vector(L)
         az -= self.env.gravity.get_value_opt(z)  # Include gravity
 
+        # Coriolis acceleration
+        _, w_earth_y, w_earth_z = self.env.earth_rotation_vector
+        ax -= 2 * (
+            -a23 * vy * w_earth_y
+            + a22 * vz * w_earth_y
+            - a33 * vy * w_earth_z
+            + a32 * vz * w_earth_z
+        )
+        ay -= 2 * (
+            a23 * vx * w_earth_y
+            - a21 * vz * w_earth_y
+            + a33 * vx * w_earth_z
+            - a31 * vz * w_earth_z
+        )
+        az -= 2 * (
+            -a22 * vx * w_earth_y
+            + a21 * vy * w_earth_y
+            - a32 * vx * w_earth_z
+            + a31 * vy * w_earth_z
+        )
+
         # Create u_dot
         u_dot = [
             vx,
@@ -1743,7 +1764,7 @@ class Flight:
         _, _, z, vx, vy, vz, e0, e1, e2, e3, omega1, omega2, omega3 = u
 
         # Create necessary vectors
-        # r = Vector([x, y, z])               # CDM position vector
+        # r = Vector([x, y, z])  # CDM position vector
         v = Vector([vx, vy, vz])  # CDM velocity vector
         e = [e0, e1, e2, e3]  # Euler parameters/quaternions
         w = Vector([omega1, omega2, omega3])  # Angular velocity vector
@@ -1896,8 +1917,9 @@ class Flight:
         # Angular velocity derivative
         w_dot = I_CM.inverse @ (T21 + (T20 ^ r_CM))
 
-        # Velocity vector derivative
-        v_dot = K @ (T20 / total_mass - (r_CM ^ w_dot))
+        # Velocity vector derivative + Coriolis acceleration
+        w_earth = Kt @ Vector(self.env.earth_rotation_vector)
+        v_dot = K @ (T20 / total_mass - (r_CM ^ w_dot)) - 2 * (w_earth ^ v)
 
         # Euler parameters derivative
         e_dot = [